{{cookiecutter.app_name}}/requirements/develop.txt
# Everything the developer needs in addition to the production requirements
-r production.txt

# Testing
pytest>=2.5.2
webtest>=2.0.15
factory-boy>=2.2.1
mock>=1.0.1

# Debug toolbar
Flask-DebugToolbar==0.9.0
